Fix a privilege escalation in `pre-receive` hook
For instance with verifying with `pre-receive` `Git Hook` the
compilation of the commit thanks to `make`.
[Current Git `pre-receive` hook](https://github.com/go-gitea/gitea/blob/5b5f7b756bfca7627b5d1326bf7fe8ef13667b7a/modules/repository/hooks.go#L20-L37) runs firstly the Gitea one named `gitea` and then the repository one named `pre-receive` (as _filename expansion in Bash is sorting alphabetically_), then verifies their exit codes. This is dangerous in my case as anyone (not having write access to the repository) is able to push a commit that will be rejected afterwards (thanks to Gitea hook) but this commit can contain a `Makefile` executing arbitrary code...
Just exiting with the Gitea hook exit code if it's non-zero solves this issue.

Note that `update` and `post-receive` should be modified in a
similar way, if they are ran even if `pre-receive` returns a non-zero
exit code.